Adds a Dockerfile and docker-compose file
First pass at starting up a blank Boston.gov install with a few
repeatable commands.

Major change is to run the "update" functions inside the hook_installs
for the bos_profile and bos_core. This is based on the assumption that
upgrading in production will not run hook_install.

Currently only tested for Boston.gov, not for The Hub.